rebild vue modules
