+++ src/frontend/src/common/heplers.js
@@ -19,6 +19,7 @@ export const createResources = (notifier) => {
     [resources.SAUCES]: new ReadOnlyApiService(resources.SAUCES, notifier),
     [resources.MISC]: new ReadOnlyApiService(resources.MISC, notifier),
     [resources.ADDRESSES]: new CrudApiService(resources.ADDRESSES, notifier),
+    [resources.ORDERS]: new CrudApiService(resources.ORDERS, notifier),
   };
 };
