+++ src/frontend/src/common/ui/index.js
@@ -1,0 +1,3 @@
+import UiRadioButton from "@/common/ui/UiRadioButton.vue";
+
+export { UiRadioButton };
