+++ src/frontend/src/main.js
@@ -4,10 +4,12 @@ import router from "@/router/index";
 import store from "@/store/index";
 import "@/plugins/vuePlugins";
 import { validator } from "@/directives";
+import { UiRadioButton } from "@/common/ui";
 
 Vue.config.productionTip = false;
 
 Vue.directive("validate", validator);
+Vue.component("UiRadioButton", UiRadioButton);
 
 new Vue({
   store,
