+++ src/frontend/src/modules/builder/SelectSauce.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="ingredients__sauce">
     <p>Основной соус:</p>
-    <RadioButton
+    <UiRadioButton
       :class="['radio', 'ingredients__input']"
       v-for="{ id, name, ...other } of items"
       :key="id"
@@ -17,19 +17,15 @@
       "
     >
       <span>{{ name }}</span>
-    </RadioButton>
+    </UiRadioButton>
   </div>
 </template>
 
 <script>
-import RadioButton from "@/common/RadioButtonNew";
 import { Builder } from "@/common/enums/entity";
 
 export default {
   name: "SelectSauce",
-  components: {
-    RadioButton,
-  },
   props: {
     items: {
       type: Array,
