+++ src/frontend/src/modules/cart/CartForm.vue
@@ -11,41 +11,58 @@
 
       <label class="input input--big-label">
         <span>Контактный телефон:</span>
-        <input type="text" name="tel" placeholder="+7 999-999-99-99" />
+        <input
+          type="text"
+          name="tel"
+          v-validate="['required']"
+          placeholder="+7 999-999-99-99"
+          :style="{
+            borderColor: currentError('tel') ? 'crimson' : 'gray',
+          }"
+        />
       </label>
 
-      <div
-        class="cart-form__address"
-        v-if="parseInt(type) === DeliveryType.NEW_ADDRESS"
+      <transition
+        name="address"
+        appear
+        enter-active-class="animate__animated animate__fadeInUp"
+        leave-active-class="animate__animated animate__fadeOutDown"
       >
-        <span class="cart-form__label">Новый адрес:</span>
+        <div
+          class="cart-form__address"
+          v-if="parseInt(type) === DeliveryType.NEW_ADDRESS"
+        >
+          <span class="cart-form__label">Новый адрес:</span>
 
-        <div class="cart-form__input">
-          <label class="input">
-            <span>Улица*</span>
-            <input type="text" name="street" />
-          </label>
-        </div>
+          <div class="cart-form__input">
+            <label class="input">
+              <span>Улица*</span>
+              <input type="text" name="street" />
+            </label>
+          </div>
 
-        <div class="cart-form__input cart-form__input--small">
-          <label class="input">
-            <span>Дом*</span>
-            <input type="text" name="house" />
-          </label>
-        </div>
+          <div class="cart-form__input cart-form__input--small">
+            <label class="input">
+              <span>Дом*</span>
+              <input type="text" name="house" />
+            </label>
+          </div>
 
-        <div class="cart-form__input cart-form__input--small">
-          <label class="input">
-            <span>Квартира</span>
-            <input type="text" name="apartment" />
-          </label>
+          <div class="cart-form__input cart-form__input--small">
+            <label class="input">
+              <span>Квартира</span>
+              <input type="text" name="apartment" />
+            </label>
+          </div>
         </div>
-      </div>
+      </transition>
     </div>
   </div>
 </template>
 
 <script>
+import { CrudCollection } from "@/common/heplers";
+
 const DeliveryType = {
   MYSELF: 1,
   NEW_ADDRESS: 2,
@@ -56,10 +73,18 @@ export default {
   data() {
     return {
       type: 1,
+      errors: [],
     };
   },
   computed: {
     DeliveryType: () => DeliveryType,
+    currentError: function () {
+      return function (id) {
+        return CrudCollection.getElement(this.errors, (item) =>
+          item.id.match(id)
+        );
+      };
+    },
   },
 };
 </script>
