--- src/frontend/src/modules/profile/index.vue
@@ -1,124 +1,0 @@
-<template>
-  <div class="layout__content">
-    <div class="layout__title">
-      <h1 class="title title--big">Мои данные</h1>
-    </div>
-
-    <!-- User info -->
-    <user-info :item="user" />
-
-    <div class="layout__address">
-      <div class="sheet address-form">
-        <div class="address-form__header">
-          <b>Адрес №1. Тест</b>
-          <div class="address-form__edit">
-            <button type="button" class="icon">
-              <span class="visually-hidden">Изменить адрес</span>
-            </button>
-          </div>
-        </div>
-        <p>Невский пр., д. 22, кв. 46</p>
-        <small>Позвоните, пожалуйста, от проходной</small>
-      </div>
-    </div>
-
-    <div class="layout__address">
-      <form
-        action="test.html"
-        method="post"
-        class="address-form address-form--opened sheet"
-      >
-        <div class="address-form__header">
-          <b>Адрес №1</b>
-        </div>
-
-        <div class="address-form__wrapper">
-          <div class="address-form__input">
-            <label class="input">
-              <span>Название адреса*</span>
-              <input
-                type="text"
-                name="addr-name"
-                placeholder="Введите название адреса"
-                required
-              />
-            </label>
-          </div>
-          <div class="address-form__input address-form__input--size--normal">
-            <label class="input">
-              <span>Улица*</span>
-              <input
-                type="text"
-                name="addr-street"
-                placeholder="Введите название улицы"
-                required
-              />
-            </label>
-          </div>
-          <div class="address-form__input address-form__input--size--small">
-            <label class="input">
-              <span>Дом*</span>
-              <input
-                type="text"
-                name="addr-house"
-                placeholder="Введите номер дома"
-                required
-              />
-            </label>
-          </div>
-          <div class="address-form__input address-form__input--size--small">
-            <label class="input">
-              <span>Квартира</span>
-              <input
-                type="text"
-                name="addr-apartment"
-                placeholder="Введите № квартиры"
-              />
-            </label>
-          </div>
-          <div class="address-form__input">
-            <label class="input">
-              <span>Комментарий</span>
-              <input
-                type="text"
-                name="addr-comment"
-                placeholder="Введите комментарий"
-              />
-            </label>
-          </div>
-        </div>
-
-        <div class="address-form__buttons">
-          <button type="button" class="button button--transparent">
-            Удалить
-          </button>
-          <button type="submit" class="button">Сохранить</button>
-        </div>
-      </form>
-    </div>
-
-    <div class="layout__button">
-      <button type="button" class="button button--border">
-        Добавить новый адрес
-      </button>
-    </div>
-  </div>
-</template>
-
-<script>
-import { mapState } from "vuex";
-import UserInfo from "@/modules/profile/UserInfo.vue";
-
-export default {
-  name: "ProfileView",
-  components: { UserInfo },
-  computed: {
-    ...mapState("Auth", ["user"]),
-  },
-  async mounted() {
-    const addresses = await this.$api.addresses.query();
-
-    console.log(addresses);
-  },
-};
-</script>
