+++ src/frontend/src/store/modules/cart.store.js
@@ -2,6 +2,7 @@ import {
   ADD_ENTITY,
   CLEAR_ENTITY,
   DELETE_ENTITY,
+  REPLACE_ENTITY,
   UPDATE_ENTITY,
 } from "@/store/mutations";
 import Module from "@/common/enums/module";
@@ -78,16 +79,22 @@ export default {
       );
     },
     clearCart({ commit }) {
-      commit(CLEAR_ENTITY, {
-        module,
-        entity: Cart.MISC,
-      });
+      commit(
+        REPLACE_ENTITY,
+        {
+          module,
+          entity: Cart.MISC,
+          payload: [],
+        },
+        { root: true }
+      );
 
       commit(
         CLEAR_ENTITY,
         {
           module,
           entity: Cart.ORDERS,
+          payload: [],
         },
         { root: true }
       );
