+++ src/frontend/src/views/Cart.vue
@@ -1,18 +1,69 @@
 <template>
-  <cart-view>
-    <template #popup>
-      <router-view />
-    </template>
-  </cart-view>
+  <form method="post" class="layout-form">
+    <slot name="popup" />
+    <main class="content cart">
+      <div class="container">
+        <div class="cart__title">
+          <h1 class="title title--big">Корзина</h1>
+        </div>
+        <div class="sheet cart__empty" v-if="isEmpty">
+          <p>В корзине нет ни одного товара</p>
+        </div>
+        <div v-else>
+          <cart-orders
+            @update="updateItem"
+            @delete="deleteItemOrder"
+            @edit="editOrder"
+            :items="orders"
+          />
+          <cart-additional
+            @add="addItem"
+            @delete="deleteItem"
+            @update="updateItem"
+            :items="misc"
+          />
+          <cart-form />
+        </div>
+      </div>
+    </main>
+    <cart-footer :sum="sumOrders" />
+  </form>
 </template>
 
 <script>
-import CartView from "@/modules/cart/CartView";
+import CartFooter from "@/modules/cart/CartFooter";
+import { mapActions, mapGetters, mapState } from "vuex";
+import CartOrders from "@/modules/cart/CartOrders";
+import CartAdditional from "@/modules/cart/CartAdditional";
+import CartForm from "@/modules/cart/CartForm";
 
 export default {
-  name: "Cart",
+  name: "CartView",
   components: {
-    CartView,
+    CartOrders,
+    CartFooter,
+    CartAdditional,
+    CartForm,
+  },
+  computed: {
+    ...mapState("Cart", ["orders"]),
+    ...mapGetters("Cart", ["isEmpty", "sumOrders"]),
+    ...mapState(["misc"]),
+  },
+  methods: {
+    ...mapActions("Cart", ["deleteItem", "updateItem", "addItem", "clearCart"]),
+    editOrder(payload) {
+      this.$store.state.Builder = payload;
+
+      this.$router.push("/");
+    },
+    deleteItemOrder(payload) {
+      this.deleteItem(payload);
+
+      if (this.orders.length < 1) {
+        this.clearCart();
+      }
+    },
   },
 };
 </script>
